moved some function to make more sense and added a display subfile
